before target network
